# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: CHORA+ API
  version: 1.0.0
  description: API REST do CHORA+ — controlo de horas e registo de atividades. Erros em application/problem+json (ADR-06).
servers:
  -
    url: "http://localhost:7071"
    description: Desenvolvimento local
...
paths:
...
  /api/v1/relatorios/horas-por-perfil:
    get:
      summary: Horas por perfil
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: contratoId
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Relatório
  /api/v1/relatorios/execucao-financeira:
    get:
      summary: Execução financeira
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: contratoId
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Relatório
  /api/v1/relatorios/ad-hoc:
    get:
      summary: "Relatórios compostos no assistente (definições, não dados)"
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Lista de receitas
  /api/v1/relatorios/ad-hoc/{id}:
    get:
      summary: Obtém a definição de um relatório ad-hoc
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Definição
        404:
          description: Inexistente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problema"
    patch:
      summary: Alterna a leitura do período entre relativa e fixa
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Definição atualizada
        403:
          description: Papel insuficiente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problema"
    delete:
      summary: Apaga um relatório ad-hoc
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Apagado
        403:
          description: Papel insuficiente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problema"
  /api/v1/relatorios/ad-hoc/{id}/executar:
    post:
      summary: Executa a receita com os dados de hoje e as permissões de quem pede
      tags:
        - Relatórios
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Tabela, ou o passo que impediu a execução"
        404:
          description: Inexistente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problema"
  /api/v1/alertas:
    get:
      summary: Lista alertas
      tags:
        - Alertas
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Lista de alertas
...
  /api/v1/auditoria:
    get:
      summary: Consulta a trilha de auditoria
      tags:
        - Auditoria
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: Eventos de auditoria
        403:
          description: Papel insuficiente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problema"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token via SDK.getAccessToken() do Azure DevOps.
  schemas:
...
    Problema:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        regra:
          type: string
        requisito:
          type: string
        base:
          type: string
        dados:
          type: object
          additionalProperties: {}
        excecaoFundamentavel:
          type: boolean
      required:
        - type
        - title
        - status
        - detail